# backend_api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart Outage MVP API",
    "description": "Backend API for outage intake, dispatch/job management, customer lookup, and audit trails.",
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "",
        "operationId": "Health",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/outages": {
      "get": {
        "tags": [
          "Outages"
        ],
        "summary": "List outages for an organization",
        "operationId": "ListOutages",
        "parameters": [
          {
            "name": "organizationId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfOutageDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Outages"
        ],
        "summary": "Create a new outage intake record",
        "operationId": "CreateOutage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOutageRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutageDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/outages/{outageId}": {
      "get": {
        "tags": [
          "Outages"
        ],
        "summary": "Get outage by id",
        "operationId": "GetOutage",
        "parameters": [
          {
            "name": "outageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutageDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/api/outages/{outageId}/status": {
      "patch": {
        "tags": [
          "Outages",
          "Audit"
        ],
        "summary": "Update outage status (with audit update)",
        "operationId": "UpdateOutageStatus",
        "parameters": [
          {
            "name": "outageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOutageStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutageDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/api/outages/{outageId}/notes": {
      "post": {
        "tags": [
          "Outages",
          "Audit"
        ],
        "summary": "Add a note to the outage timeline",
        "operationId": "AddOutageNote",
        "parameters": [
          {
            "name": "outageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddOutageNoteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutageUpdateDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/api/outages/{outageId}/updates": {
      "get": {
        "tags": [
          "Outages",
          "Audit"
        ],
        "summary": "List outage timeline updates",
        "operationId": "ListOutageUpdates",
        "parameters": [
          {
            "name": "outageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfOutageUpdateDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "List jobs for an organization",
        "operationId": "ListJobs",
        "parameters": [
          {
            "name": "organizationId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "outageId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfJobDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Create a new job for an outage",
        "operationId": "CreateJob",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{jobId}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get job by id",
        "operationId": "GetJob",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/api/jobs/{jobId}/assign": {
      "patch": {
        "tags": [
          "Jobs",
          "Dispatch",
          "Audit"
        ],
        "summary": "Assign job to crew/user (transitions to assigned)",
        "operationId": "AssignJob",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignJobRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/api/jobs/{jobId}/status": {
      "patch": {
        "tags": [
          "Jobs",
          "Audit"
        ],
        "summary": "Update job status (writes job_status_events audit)",
        "operationId": "UpdateJobStatus",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/api/jobs/{jobId}/events": {
      "get": {
        "tags": [
          "Jobs",
          "Audit"
        ],
        "summary": "List job status audit events",
        "operationId": "ListJobStatusEvents",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfJobStatusEventDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/customers/{customerUserId}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Get a customer by user id",
        "operationId": "GetCustomer",
        "parameters": [
          {
            "name": "customerUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/api/customers": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Search customers by name/email/phone",
        "operationId": "SearchCustomers",
        "parameters": [
          {
            "name": "organizationId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfCustomerDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LocationDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "addressLine1",
          "city",
          "country"
        ],
        "properties": {
          "addressLine1": {
            "type": "string"
          },
          "addressLine2": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          }
        }
      },
      "CustomerDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "email",
          "fullName",
          "role",
          "isActive"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "OutageDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "organizationId",
          "locationId",
          "title",
          "severity",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": { "type": "string" },
          "organizationId": { "type": "string" },
          "reportedByUserId": { "type": "string", "nullable": true },
          "customerUserId": { "type": "string", "nullable": true },
          "locationId": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "severity": { "type": "string" },
          "status": { "type": "string" },
          "startedAt": { "type": "string", "format": "date-time", "nullable": true },
          "resolvedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "JobDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "organizationId",
          "outageId",
          "status",
          "priority",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": { "type": "string" },
          "organizationId": { "type": "string" },
          "outageId": { "type": "string" },
          "assignedCrewId": { "type": "string", "nullable": true },
          "assignedToUserId": { "type": "string", "nullable": true },
          "status": { "type": "string" },
          "priority": { "type": "string" },
          "notes": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "JobStatusEventDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "jobId",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "jobId": { "type": "string" },
          "status": { "type": "string" },
          "changedByUserId": { "type": "string", "nullable": true },
          "message": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "OutageUpdateDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "outageId",
          "updateType",
          "message",
          "createdAt"
        ],
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "outageId": { "type": "string" },
          "updateType": { "type": "string" },
          "message": { "type": "string" },
          "createdByUserId": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateOutageRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "organizationId",
          "title",
          "severity",
          "location"
        ],
        "properties": {
          "organizationId": { "type": "string" },
          "reportedByUserId": { "type": "string", "nullable": true },
          "customerUserId": { "type": "string", "nullable": true },
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "severity": { "type": "string" },
          "startedAt": { "type": "string", "format": "date-time", "nullable": true },
          "location": { "$ref": "#/components/schemas/LocationDto" }
        }
      },
      "UpdateOutageStatusRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status"
        ],
        "properties": {
          "status": { "type": "string" },
          "message": { "type": "string", "nullable": true },
          "changedByUserId": { "type": "string", "nullable": true },
          "resolvedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "AddOutageNoteRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "message"
        ],
        "properties": {
          "message": { "type": "string" },
          "createdByUserId": { "type": "string", "nullable": true }
        }
      },
      "CreateJobRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "organizationId",
          "outageId",
          "priority"
        ],
        "properties": {
          "organizationId": { "type": "string" },
          "outageId": { "type": "string" },
          "priority": { "type": "string" },
          "notes": { "type": "string", "nullable": true }
        }
      },
      "AssignJobRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "assignedCrewId": { "type": "string", "nullable": true },
          "assignedToUserId": { "type": "string", "nullable": true },
          "message": { "type": "string", "nullable": true },
          "changedByUserId": { "type": "string", "nullable": true }
        }
      },
      "UpdateJobStatusRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status"
        ],
        "properties": {
          "status": { "type": "string" },
          "message": { "type": "string", "nullable": true },
          "changedByUserId": { "type": "string", "nullable": true }
        }
      },
      "PagedResultOfOutageDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "limit",
          "offset",
          "returned"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OutageDto"
            }
          },
          "limit": { "type": "integer", "format": "int32" },
          "offset": { "type": "integer", "format": "int32" },
          "returned": { "type": "integer", "format": "int32" }
        }
      },
      "PagedResultOfJobDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "limit",
          "offset",
          "returned"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobDto"
            }
          },
          "limit": { "type": "integer", "format": "int32" },
          "offset": { "type": "integer", "format": "int32" },
          "returned": { "type": "integer", "format": "int32" }
        }
      },
      "PagedResultOfCustomerDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "limit",
          "offset",
          "returned"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerDto"
            }
          },
          "limit": { "type": "integer", "format": "int32" },
          "offset": { "type": "integer", "format": "int32" },
          "returned": { "type": "integer", "format": "int32" }
        }
      },
      "PagedResultOfOutageUpdateDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "limit",
          "offset",
          "returned"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OutageUpdateDto"
            }
          },
          "limit": { "type": "integer", "format": "int32" },
          "offset": { "type": "integer", "format": "int32" },
          "returned": { "type": "integer", "format": "int32" }
        }
      }
    }
  }
}